refactor: split unix-specific tests
